pyproject.toml
[tool.poetry]
name = "ai-pbpk-mpc"
version = "0.1.0"
description = "End-to-end AI-assisted PBPK simulation with MPC and reporting"
authors = ["AI-PBPK Team <devnull@example.com>"]
readme = "README.md"
packages = [{ include = "ai_pbpk" }]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pandas = "^2.2.2"
scipy = ">=1.11,<1.13"
pydantic = "^2.7.4"
PyYAML = "^6.0.2"
typer = "^0.12.5"
scikit-learn = "^1.5.2"
xgboost = "^2.1.1"
joblib = "^1.4.2"
casadi = "^3.6.5"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
plotly = "^5.22.0"
mlflow = "^2.14.1"
rich = "^13.7.1"
loguru = "^0.7.2"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.6", extras=["standard"]}
pyarrow = "^17.0.0"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=ai_pbpk --cov-report=term-missing"
testpaths = [
  "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


